Move `RemoteFeedLoader` components to production
As we move files from a different target (production), we have
to decide whether we want to expose internal types (using `@testable`)
or make the necessary parts `public`. The decision was to make them
`public` because we can have more fine-grained control over what to
expose and what shouldn’t. The `public` part is exposed because
other modules can use it. `RemoteFeedLoader` was made `final` because
we don’t expect any other type to inherit from it.