Guarantee that we can load more than once
It is crucial to ensure that the behavior is only executed once if
we expect that. To ensure it, we should record all the calls to our
spy, and one way is to record the URLs we receive. This way, we can not
only assert the number of API calls but also check the value
of the parameters passed.