Replace optional params with enum with success/failure cases
Initially, the `get` method of `HTTPClient` received an optional
Error and an optional HTTPURLResponse. However, in reality, only one
can have value and the cases of both of them have value, or none
of them are invalid.

It is a better option to represent this with a type and the proper
type for this is an enum.