Add `RemoteFeedLoader.Result` type with success/failure cases
While exploring the happy path, we realized we needed to change
the production interface. To do the refactoring, we commented out
the test and did the interface changes, backed by the unit tests.

The `load` method should return with a `[FeedItem]` type on success.
To avoid the explosion of invalid cases, we are using an enum
to represent the success and failure cases instead of having
two separate parameters (feed items and error). This simplifies the
receiving code and eliminates the invalid cases (like both error and
data are set, or none of them are set).