Move `Decodable` logic to a new private `Item` struct to decouple the Feed Feature module from API implementation details.
Previously, the `FeedItem` struct implements the `Decodable` protocol.
Because of the differences in the JSON field `image` and the `imageURL`
property of the `FeedItem`, we needed to add `CodingKeys` to map between
the two. However, it couples the Feed Feature module to the remote API,
so we must update multiple modules when the backend API changes.
Furthermore, a type can conform to the `Decodable` protocol only once,
therefore, we cannot conform to different backend APIs with the
`FeedItem` type.

Adding an internal type that conforms to the backend API and can be
mapped to `FeedItem` solves the problem. It allows us to follow API
changes and even support multiple backends.