Move all mapping logic to the `FeedItemsMapper`
Extracting the logic from `RemoteFeedLoader` to a private helper method
created a retain cycle in `load`. To avoid the retain cycle, we made
the helper method static. It turned out that the helper method has a
better place in `FeedItemMapper`, and it allowed us to refactor
the logic by merging it with the previous implementation of `map`.